rsaenh: Use Interlocked functions for reference counting.

+++ dlls/rsaenh/handle.h
@@ -36,7 +36,7 @@ typedef void (*DESTRUCTOR)(OBJECTHDR *object);
 struct tagOBJECTHDR
 {
     DWORD       dwType;
-    UINT        refcount;
+    LONG        refcount;
     DESTRUCTOR  destructor;
 };
